refactor: flatten some `me`-related pages components logic
