--- src/components/Dashboard/ProposalsTable/index.tsx
@@ -1,1 +1,0 @@
-export * from "./ProposalsTable";
